requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic-settings
pytest
pytest-asyncio
httpx
reportlab
